add useFilter composable to integrate easily with SearchFilter component
